pyproject.toml
[tool.poetry]
name = "multimodal-recommendation"
version = "0.1.0"
description = ""
authors = ["Ryan Monks"]
readme = "README.md"
packages = [{include = "multimodal_recommendation"}]

[tool.poetry.dependencies]
python = "3.11.5"
transformers = "^4.35.0"
pandas = "^2.1.2"
torch = "^2.1.0"
pillow = "^10.1.0"
sentence-transformers = "^2.2.2"
langchain = "^0.0.334"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
